pages: pluralise group, id, application links
