+++ content/applications/~haddef-sigwen/dummy.md
@@ -1,0 +1,11 @@
++++
+title = "Dummy"
+shortcode = "~haddef-sigwen/dummy"
+description = "All the features you'd want from your Urbit."
+license = "MIT"
+image = "https://fabled-faster.nyc3.digitaloceanspaces.com/fabled-faster/2021.4.02..21.52.41-UC.png"
+developer = "Haddef Sigwen LLC."
+website = "https://haddef-sigwen.arvo.network"
++++
+
+Dummy provides _real_ Lorem Ipsum content.
