+++ pages/grants/[slug].js
@@ -44,7 +44,7 @@ export default function Grant({ post, markdown, search }) {
             </div>
           ) : null}
           {post.extra.ship ? (
-            <Link href={`/id/${post.extra.ship}`}>
+            <Link href={`/ids/${post.extra.ship}`}>
               <a className="type-sub-bold text-wall-500 font-mono">
                 {post.extra.ship}
               </a>
