+++ pages/groups/[[...group]].js
@@ -2,6 +2,7 @@ import Head from "next/head";
 import Link from "next/link";
 import { getPage } from "../../lib/lib";
 import { join } from "path";
+import Meta from "../../components/Meta";
 import CopyLink from "../../components/CopyLink";
 import Container from "../../components/Container";
 import SingleColumn from "../../components/SingleColumn";
@@ -14,12 +15,12 @@ import MetadataBlock from "../../components/gateway/MetadataBlock";
 import MetadataLink from "../../components/gateway/MetadataLink";
 import Description from "../../components/gateway/Description";
 
-const ApplicationPage = ({ data, markdown, params }) => {
-  const { application } = params;
-  if (!ob.isValidPatp(application[0])) {
-    return <Gateway404 type="application" />;
+const GroupPage = ({ data, markdown, params }) => {
+  const { group } = params;
+  if (!ob.isValidPatp(group[0])) {
+    return <Gateway404 type="group" />;
   }
-  const image = `https://urbit-id-og-cards-7mdrgkd36-urbit.vercel.app/${
+  const image = `https://urbit-id-og-cards-ltjw3771z-urbit.vercel.app/${
     data.title
   }?images=${
     data.image ? data.image : "https://media.urbit.org/logo/urbit-logo-card.png"
@@ -28,7 +29,7 @@ const ApplicationPage = ({ data, markdown, params }) => {
   return (
     <Container>
       <Head>
-        <title>{data.title} • Applications • urbit.org</title>
+        <title>{data.title} • Groups • urbit.org</title>
         <link rel="icon" type="image/png" href="/images/favicon.ico" />
         <meta
           name="twitter:card"
@@ -37,14 +38,13 @@ const ApplicationPage = ({ data, markdown, params }) => {
         />
         <meta
           name="og:title"
-          content={`${data.title} • Applications • urbit.org`}
+          content={`${data.title} • Groups • urbit.org`}
           key="title"
         />
         <meta
           name="og:description"
           content={
-            data?.description ||
-            "View more about this application on urbit.org."
+            data?.description || "View more about this group on urbit.org."
           }
           key="description"
         />
@@ -55,25 +55,23 @@ const ApplicationPage = ({ data, markdown, params }) => {
           <GatewayHeader
             title={data.title}
             image={data?.image}
-            item="Urbit Application"
+            item="Urbit ID"
           />
           <div className="flex flex-wrap md:flex-nowrap justify-between">
             <MetadataBlock
-              title="License"
-              content={data.license ? data.license : "Unknown"}
+              title="Group Type"
+              content={data.type ? data.type : "Unknown"}
             />
-            {data?.website && (
-              <MetadataLink
-                title="Website"
-                href={data?.website}
-                content={
-                  data?.website ? data.website.replace(/^https?:\/\//, "") : ""
-                }
-              />
-            )}
             <MetadataBlock
-              title="Developer"
-              content={data.developer ? data.developer : "Unknown"}
+              title="Members"
+              content={
+                data.participant_range ? data.participant_range : "Unknown"
+              }
+            />
+            <MetadataLink
+              title="Host"
+              href={`/ids/${group[0]}`}
+              content={group[0]}
             />
             <CopyLink
               className="basis-1/2 md:basis-auto"
@@ -82,7 +80,7 @@ const ApplicationPage = ({ data, markdown, params }) => {
           </div>
           <Description
             description={data.description}
-            fallback="An application on Urbit."
+            fallback="A group on Urbit."
             markdown={markdown}
           />
           <hr className="text-wall-200" />
@@ -92,23 +90,20 @@ const ApplicationPage = ({ data, markdown, params }) => {
             </h3>
             <div className="flex flex-col space-y-4">
               <p className="text-sm font-semibold text-wall-400">
-                Learn how to install an Urbit application
+                Get on the network and join this group.
               </p>
-              <Link href="/application/guide">
+              <Link href="/groups/guide">
                 <button className="button-lg max-w-xs bg-green-400 text-white">
-                  Installing Urbit applications
+                  How to join a group
                 </button>
               </Link>
             </div>
           </div>
           <hr className="text-wall-200" />
           <div className="flex flex-col space-y-1">
-            <p>
-              Have an application you'd like to share publicly through
-              urbit.org?
-            </p>
-            <Link href="/application/submit">
-              <a className="type-ui text-green-400">Submit your application</a>
+            <p>Have a group you'd like to share publicly through urbit.org?</p>
+            <Link href="/groups/submit">
+              <a className="type-ui text-green-400">Submit your group</a>
             </Link>
           </div>
           <Link href="/">
@@ -127,18 +122,11 @@ export const getServerSideProps = async ({ params, res }) => {
   );
 
   let { data, content } = getPage(
-    join(
-      process.cwd(),
-      "content/application",
-      params.application?.join("/") || "/"
-    )
+    join(process.cwd(), "content/groups", params.group?.join("/") || "/")
   ) || { data: {}, content: "" };
 
   if (!data.title) {
-    data = {
-      title: params.application?.join("/"),
-      description: "An application on Urbit.",
-    };
+    data = { title: params.group?.join("/"), description: "A group on Urbit." };
   }
 
   const markdown = await Markdown({ post: { content: content } }, true);
@@ -152,4 +140,4 @@ export const getServerSideProps = async ({ params, res }) => {
   };
 };
 
-export default ApplicationPage;
+export default GroupPage;
