+++ pages/updates/[slug].js
@@ -48,7 +48,7 @@ export default function Post({
             </div>
           ) : null}
           {post?.extra?.ship ? (
-            <Link href={`/id/${post.extra.ship}`}>
+            <Link href={`/ids/${post.extra.ship}`}>
               <a className="type-sub-bold text-wall-500 font-mono">
                 {post?.extra?.ship}
               </a>
